fix custom contest
