+++ src/features/custom_contests/components/Form/ProblemStep.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useCallback, Suspense } from "react";
+import React, { useCallback, Suspense } from "react";
 import Stack from "@mui/material/Stack";
 import Box from "@mui/material/Box";
 import KeyboardDoubleArrowLeftIcon from "@mui/icons-material/KeyboardDoubleArrowLeft";
@@ -10,13 +10,12 @@ import { Problem, Tag } from "@features/problems/problem";
 import { useFetchProblems } from "@features/problems/hooks/useFetchProblem";
 import { NumberOfProblems } from "@features/custom_contests/components/Form/NumberOfProblems";
 import { ProblemsTag } from "@features/custom_contests/components/Form/ProblemsTag";
-import { ExpectedParticipants } from "@features/custom_contests/components/Form/ExpectedParticipants";
+import { ExcludeSolved } from "@features/custom_contests/components/Form/ExcludeSolved";
 import { Difficulty } from "@features/custom_contests/components/Form/Difficulty";
 import { ErrorMessage } from "@features/ui/component/ErrorMessage";
 import { SelectedProblemsTable } from "./SelectedProblemsTable";
-import { useFetchExpectedParticipantsSolvedProblems } from "@features/custom_contests/hooks/useFetchSubmissions";
-import { getProblemKey } from "@features/problems/utils";
 import { CircularProgress } from "@features/ui/component/CircularProgress";
+import { SelectProblems } from "./SelectProblems";
 
 type Props = {
   setActiveStep(step: number): void;
@@ -35,41 +34,6 @@ export const ProblemStep: React.FC<Props> = ({
 }) => {
   const { data } = useFetchProblems();
 
-  // const expectedParticipants =
-  //   getValues().problemsFilter.expectedParticipants.map((v) => v.name);
-  // const [shouldFetch, setShouldFetch] = useState(false);
-  // const { solvedSet } = useFetchExpectedParticipantsSolvedProblems(
-  //   expectedParticipants,
-  //   shouldFetch
-  // );
-  // useEffect(() => {
-  //   setShouldFetch(false);
-  // }, [solvedSet]);
-
-  const selectProblems = useCallback(
-    (problems: Problem[]) => {
-      const values = getValues();
-      return problems
-        .filter(
-          (problem) =>
-            (problem.rating ?? 0) >=
-              (values.problemsFilter.difficultyFrom ?? 0) &&
-            (problem.rating ?? 0) <=
-              (values.problemsFilter.difficultyTo ?? 0) &&
-            values.problemsFilter.includeTags.every((includeTag) =>
-              (problem.tags as Tag[]).includes(includeTag)
-            ) &&
-            values.problemsFilter.excludeTags.every(
-              (excludeTag) => !(problem.tags as Tag[]).includes(excludeTag)
-            )
-        )
-        .sort(() => Math.random() - 0.5)
-        .slice(0, values.problemsFilter.count ?? 0)
-        .sort((a, b) => (a.rating ?? 0) - (b.rating ?? 0));
-    },
-    [getValues]
-  );
-
   return (
     <Box pb={{ xs: 2, md: 4 }}>
       <Box sx={{ px: { xs: 1, md: 4 }, py: 3 }}>
@@ -77,26 +41,19 @@ export const ProblemStep: React.FC<Props> = ({
           <NumberOfProblems control={control} errors={errors} />
           <Difficulty control={control} errors={errors} />
           <ProblemsTag control={control} errors={errors} />
-          <ExpectedParticipants control={control} errors={errors} />
-        </Stack>
-        <Stack direction="row" justifyContent="flex-end" sx={{ my: 2 }}>
-          <Button
-            onClick={() => {
-              // setShouldFetch(true);
-              data && setValue("problems", selectProblems(data));
-            }}
-            disabled={!data}
-          >
-            Generate Problems
-          </Button>
+          <ExcludeSolved
+            control={control}
+            errors={errors}
+            getValues={getValues}
+          />
         </Stack>
         <Suspense fallback={<CircularProgress />}>
-          <Controller
-            name="problems"
+          <SelectProblems
+            data={data}
             control={control}
-            render={({ field }) => (
-              <SelectedProblemsTable isEdit={true} field={field} />
-            )}
+            errors={errors}
+            getValues={getValues}
+            setValue={setValue}
           />
         </Suspense>
         {errors.problems && <ErrorMessage message={errors.problems.message} />}
